Automation Updates
Why these changes are being introduced:
The ECR infrastructure for this application has already been deployed.
The outputs from that Terraform run are provided here for the Makefile
and the three GitHub Action workflows for automating the container
publishing process to our AWS Organization.

How this addresses that need:
* Add Terraform-generated content to the Makefile to allow the
developer to easily build and push containers to the Dev1 ECR
repository from the local CLI
* Create the dev-build workflow (from the Terraform output) for
automated container builds & uploads to Dev1 on pushes to any feature
branch
* Create the stage-build workflow (from the Terraform output) for
automated container builds & uploads to Stage-Workloads on any merges
to the main branch
* Create the prod-promote workflow (from the Terraform output) for
automated container uploads to Prod-Workloads on any tagged releases on
the main branch

Side effects of this change:
Once this is merged to the `main` branch, automated deployments will be
live.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-656
* https://mitlibraries.atlassian.net/browse/IN-657