check multiple address types
Why these changes are being introduced:

We want to use the address specified as the payment address in alma.
An address can have multiple types in alma. We were only checking the first type of each
address to see if it was a payment address, so in these cases we were not correctly
identifying the payment address

How this addresses that need:

Checks all address types for each address to determine if an address is the payment
address